moved equipmentList, firearms, melee, tables and uniformAndArmourTypes to data folder
