+++ src/data/equipmentList.js
