+++ src/data/firearms.js
