+++ src/data/firearms.spec.js
