+++ src/data/uniformAndArmourTypes.js
